store: remove handling of gitignore in diff
Summary:
EdenFS behavior appears to be slightly different from Mercurial (and I presume
git) when it comes to files ignored by a gitignore but known to source control.
In particular, Mercurial appear to not look at the gitignore when comparing 2
commits, or when comparing the working copy to another commit.

For instance, if we try to manually reproduce the
diff_added_tree_top_level_not_materialized_with_gitignore, here is what we see:
```
% hg status --rev 4694c115b9bba800b6d2cf47e10f556f3e8e5b7e
A root/src/foo/r.txt
% eden debug thrift getScmStatus "/data/users/xavierd/fbsource" False 4694c115b9bba800b6d2cf47e10f556f3e8e5b7e
ScmStatus(
    entries={},
    errors={})
```

That is: EdenFS doesn't report any changes, while Mercurial does.

Fixing this issue is just a matter of removing all the gitignore handling code
from store/Diff.cpp and fixing the callers.

Reviewed By: quark-zju

Differential Revision: D48705684

fbshipit-source-id: 5526e28fe3c7b3a329119a245c09b1fc11372e97