UncommittedChanges: show chunk selection inline
Summary:
The tooltip is hard to layout:
{F1083502634}

Showing the chunk selections inline so we don't
have to worry much about the height.

Right now only the unified diff view is used. There is no editing or expansion.
This UI seems straightforward to understand.

I thought about switching the tooltip to a modal before. But that is annoying,
for example, the modal has to duplicate operations like commit/amend/discard to
avoid the "double confirm" issue. There is no way to use the right panel for
structured commit message editing. By showing the chunk selector inline we avoid
button duplication and support reusing the right panel.

Reviewed By: evangrayk

Differential Revision: D48970672

fbshipit-source-id: a08b8d57077d3a5dc9c13fd409b8f3c04dcbdddb