use promise for fetching shelved state
Summary: Instead of using an atom to track state for shelved changes, use a promise, which we refresh each time you load the shelved changes list. This way, we fetch a new list of changes each time. It's still possible for the list of shelves to change while the dropdown is open, but I think that's not very common. Perhaps we should refetch periodically while it's open...in the future.

Reviewed By: quark-zju

Differential Revision: D48992198

fbshipit-source-id: abc6f26752948aa7aa68e3c2dade90fe7c24da61