Add unshelve operation
Summary:
Add operation to run unshelve. The optimistic applier needs to add the same file changes to the list of uncommitted changes, and de-dupe with existing uncommitted changes.

Sort of like uncommit, but slightly simpler because we don't have to change any commits optimistically.

Also use OperationDisabledButton to get a spinner when the operation is running.

I think there's still more work to do on the UI when running unshelve, since we should be hiding the shelve once you've unshelved it.

Reviewed By: quark-zju

Differential Revision: D48982813

fbshipit-source-id: c00646f3fb77a09afa72fca50d2b7e945f921d36