Add a way to fetch shelved changes
Summary:
Add api to fetch shelved changes, and call to `sl log` that can fetch the info for shelves. This needs to use a different template than the one we use for `CommitInfo`s. We need to fetch additional fields like `{shelvename}`, and we don't need all the fields from `CommitInfo`, like `predecessors`.

I duplicated the way we define the fields for this. There may be a way to consolidate these into a util or something so we don't need any duplication, but I'll put that off until we have yet another format of commit template we need to fetch.

The fetch works because shelves are just commits, and we can use the `shelved()` revset to find them all.

Reviewed By: quark-zju

Differential Revision: D48982809

fbshipit-source-id: 9e2e9fdf946794bace01dc7d95958032acfc37ef