Make GitDeltaManifest derivable
Summary: This diff includes all the boilerplate code that goes into making a type derivable. Since `GitDeltaManifest` is a `Manifest` type, I have followed the pattern of having a `Root` type (that points to the actual manifest) as the derivable type instead of the core manifest type itself. In this case, the root type is `RootGitDeltaManifestId` which contains the id of the `GitDeltaManifest` for a given commit. The actual logic for deriving `GitDeltaManifest` would be added in the next diff.

Reviewed By: markbt

Differential Revision: D48756522

fbshipit-source-id: 847780f7f51a78bc644c4b6f7791765e3877dd57