Use MPath instead of Option<NonRootMPath>
Summary: The third diff in the series of replacing `Option<NonRootMPath>` usages with `MPath`. This is much bigger than the previous diffs because I have updated all manifest type ops and derivation code to use `MPaths` and since pretty much everything in Mononoke uses these types, I had to make a lot of associated changes.  After this, I will be updating the `MononokePath` implementation to basically be an alias for `MPath` and making similar changes in `RepoPath`. However, this diff completes the pre-req for doing Git work so further progress might not be immediate.

Reviewed By: markbt

Differential Revision: D49014566

fbshipit-source-id: 0f1d902a6663a73e7fd353a248a91acd06d73bd7