Introduce GitDeltaManifest type
Summary: With all the dependencies taken care of, its time to introduce the `GitDeltaManifest` type as a `ShardedMap` type. Most of the code in this diff is `Blobstore` and `Thrift` boilerplate so that we can use `GitDeltaManifest` as a `ShardedMap` type addressed by `GitDeltaManifestId`. The `GitDeltaManifest` blobs would be stored with key `repoxxx.gdm.blake2.{BLAKE2_HASH_OF_COMMIT}`

Reviewed By: markbt

Differential Revision: D48736892

fbshipit-source-id: c364f1e998518650af6f27911748f7f7c192b402