Provide override for test repo factory's CoreContext
Summary:
## What
Add a `CoreContext` field and `with_core_context` associated function to `TestRepoFactory`.

## Why
In the next diff (D48829057) I'll use `TestRepoFactory` to create a temporary mononoke repo for the partial repository.

If I don't do this, the log level options I pass to the binary will not be used in the test repo's code, so I can't use the tool in the integration tests...

Differential Revision: D48946892

fbshipit-source-id: 7066bcbf6b4252b047b435bb1752848fd93b2a63