turn on autocargo for edenscmnative/bindings/modules
Summary:
There are a bunch of crates in here with manually maintained Cargo.tomls. We just had a build breakage due to unsatisfiable "crossbream" dependency, so let's try switching these to autocargo.

We still don't want autocargo for edenscmnative/bindings, so I played with the include/exclude globs to (hopefully) be equivalent to before but with the addition of edenscmnative/bindings/modules.

Reviewed By: sggutier

Differential Revision: D48832718

fbshipit-source-id: adba1552818f052006994f9ba26244fbfae47d92