sharded_map: implement a way to create a sharded map node from independent values and other sharded map nodes
Summary: In order to optimize derivation of sharded manifests, this diff implements a function to construct a new sharded map node out of a map of values and other sharded map nodes. The key for every input sharded map node is a prefix that's prepended to it, which represents that all keys having this prefix is entirely represented by this sharded map node. This function returns an error if the key for an input sharded map node is a prefix of any other key in the input map.

Differential Revision: D48954392

fbshipit-source-id: 4a6f549cd2501ed9ea89f889dae808a6f054bbe1