Simplify FaultInjector::check and checkAsync
Summary:
We have multiple FaultInjector::check* templates with different parameters,
which will lead to an unweildy number of implementations when we add one that
returns a Try.  This cleans things up by overloading constructKey to forward
std::string_view directly to the implementations, so we can then remove the
versions that accept keyValue as a string_view.

Reviewed By: chadaustin

Differential Revision: D49114096

fbshipit-source-id: 1b25ea365537e285c65053468a48e837b983d5ab