python-modules: compiled Python modules
Summary:
One of the pain point of buck build is that it builds a "par", then unzips it,
then use a customized bash script to let hg_rust know where to find the
modules. It is kind of hacky (ex. the "par" work is a waste of time unzipping,
the bash script means no Windows support), and requires workarounds like
D38718449.

It'd be nice if the compiled Rust binary contains the edenscm Python modules
so it can load them inline instead of requiring a specific filesystem layout
or special environment variables.

This diff is a first step to get edenscm/ pure Python modules compiled so
they can be embedded later.

Right now it only includes first-party modules. In the future, we might want
to include Python's stdlib on Windows to avoid slow file reading, and migrate
pytz to 3.10 stdlib zoneinfo to drop the pytz dependency.

Reviewed By: muirdm

Differential Revision: D49149355

fbshipit-source-id: 215f3b1fe8267eb521d0203515a7839571d52fdf