Don't hide blank commit on the right, add details to empty state
Summary:
Showing the blank right commit makes it more clear that there's a place that your changes will move to, and makes the mental model of split more clear on first load, that there's a set of  commits laid out horizontally which you can move lines of code throughout.

This way, we also show `1/2` for the first commit by default, hinting that there are more commits that you'll be making.

Reviewed By: muirdm

Differential Revision: D49159468

fbshipit-source-id: 41880b29416cef7ad15a89d0638d7c8e19feed1d