Allow editing any commit's title in split
Summary:
Previously, we made commits with multi-line descriptions non-editable in split. I think it's more useful to let you change the title, and just keep whatever the existing description is.

In the future, we should probably also show you the description somehow (on hover?).

Here, I make the editing operation on the commit title field only affect the title, which gets split out for the editor.

Making every commit title editable is also more consistent for the UI, so it's not like, "why is only this commit's title not editable?"

Reviewed By: muirdm

Differential Revision: D49161482

fbshipit-source-id: 8157577c7d704cd7e44ec6948075ff294f0f920f