Don't select empty buttons
Summary:
Before, we used `' '` to ensure alignment. But now that we have a table, it's sufficient to have an empty `<td />`. So make the gutters null when they don't have an arrow in them.

This also fixes user selection to not include the gutters, since the `' '` was selectable, but now it's not

Reviewed By: muirdm

Differential Revision: D49156738

fbshipit-source-id: 5a4aed9388f3bbbd8122323573c8c807d4e514df