Show expansion hint on hover of context lines
Summary:
I was considering changing the ~~~~ context expander to match the comparison view, but I found that I didn't like how it looked. So let's keep the ~~~~. One small thing we can do is make it more clear that clicking will expand, by showing some text on hover.

Admittedly, by using CSS for this, it's not translatable.

Reviewed By: muirdm

Differential Revision: D49156741

fbshipit-source-id: d7229b439b721f72054b432cf8102da9d85c5ef3