Hide open file button in file header
Summary:
Since we re-used the file header from splitdiffview, we had an "open file" button next to each file. We don't want to take you out of the split commit UI, so we hide this button.

To do this, we can expose a new UI which lets you replace the open file button with arbitrary other content. We'll use this later to add left/right buttons for the entire file

Reviewed By: muirdm

Differential Revision: D49156739

fbshipit-source-id: 96e5899f06f65669bf6715e93736489cbc213a70