merge: log command for merge conflicts
Summary:
I'd like to differentiate the rebase operations trigerred by `amend`
and `rebase`.

For example, if a user deleted new added changes from an early commit of a
stack and then amend it. We cannot tell if the user wanted to delete the change
or move the change to a later commit.

Reviewed By: muirdm

Differential Revision: D49157532

fbshipit-source-id: 1221e1a0191cbd77f3cdd2b1f871443f3202621c