fix language name classifier
Summary:
`FilenameClassifier.getDisplayNameForLanguageId` takes a language ID like `hack` and is supposed to turn it into a display name like `Hack`.

first, it has to turn the language ID into a textmate scope name like `source.hack`, so that it can look up the `source.hack` textmate config.

but it was using `findScopeNameForAlias` instead of `findScopeNameForLanguage` to map the language ID to scope name.

we already have the language ID, not an alias, so the indirection through aliases is unnecessary -- and in Hack's case it's broken[1].

so this diff makes `getDisplayNameForLanguageId` first treat its arg as a language id, and fall back to the alias.

this is the same approach that highlighting uses:

https://www.internalfb.com/code/fbsource/[f8f30ba6a489]/xplat/vscode/vscode-extensions/code-compose/src/chat/TextMateMarkdownRenderer.ts?lines=156-160

[1] for some reason, the `ini` config seems to define an alias for Hack Configuration (ok) but also `hack`. this causes `findScopeNameForAlias('hack')` to return `'ini'`! the textmate grammar should be fixed, but that's a different bug.

Reviewed By: bolinfest

Differential Revision: D49191522

fbshipit-source-id: b6c0844dbf8ad2ea40ea3411037482f40db4679a