Button to insert blank commit
Summary:
Add a button between columns in the split view to add new blank commits. This allows you to decide you want finer granularity splitting midway through, and not have to manually shift everything over to make room for a new commit.

One TODO:
The way this works, it's creating "real" commits as opposed to the fake commit added to the end of the split. So if you make an empty commit and run split, it will actually make an empty commit.

I guess we should filter out empty commits when we go to apply the stack edit...?

What's more, it seems debugexportstack doesn't like empty commits, it fails to load split after creating some empty commits in my stack

Reviewed By: quark-zju

Differential Revision: D49193835

fbshipit-source-id: 787c21e8f23ec309efb9fee76e8e6294eb941ca3