Introduce DeltaInstructionChunk and DeltaInstructionChunkId
Summary:
The raw `GitDeltaInstructions` can be potentially quite large (half a GB in some cases). If these delta instructions are stored inline with the `GitDeltaManifestEntry` then the overall size of the `GitDeltaManifest` would become way too large even if its chunked. Instead of storing the actual instructions directly, we store the total number of chunks that would be created for storing the instructions. Note that we don't store the list of Chunks Ids anywhere. The IDs of the chunk can be determined based on the hashing scheme used. The ID is the hash of the actual_commit + actual_mpath + base_commit + base_mpath + index of the chunk. Except the index, all other information is available in `GitDeltaManifestEntry` and thus can be used to access all the chunks.

The follow up diff will contain the actual chunking logic for splitting up the blobs.

Differential Revision: D49098610

fbshipit-source-id: 5eeacb3913c45f07193def417a7471cdd836bb70