testing: improve compatibility with doctest on path
Summary:
Usually the doctests are run using module names (doctest:module.name). But we
also support file paths for convenience. Update the logic to be aware of the
new "<static>" modules.

Force loading static (not on-disk) modules, then check the doctest runs with
on-disk file path:

  ~/hg/edenscm % export SAPLING_PYTHON_HOME=
  ~/hg/edenscm % lrt changelog.py
  # Ran 1 tests, 0 skipped, 0 failed.

Check if the module source is changed, there is a warning:

  ~/hg/edenscm % echo >> changelog.py
  ~/hg/edenscm % lrt changelog.py
  warning: doctest is using an older <static> version of module edenscm.changelog that no longer matches on-disk .../edenscm/changelog.py
  # Ran 1 tests, 0 skipped, 0 failed.

Check that the error includes 'doctest:' hint when the module is not found:

  ~/hg/edenscm % touch x.py
  ~/hg/edenscm % lrt x.py
  ...
  RuntimeError: cannot find Python module name for path='.../scm/edenscm/x.py' to run doctest
  try 'doctest:module.name' instead of file path

Reviewed By: zzl0

Differential Revision: D49238721

fbshipit-source-id: 933a918ffe87f7d969a8e88b8f471a7383d2c07b