git: update "date" to be max (committer date, author date)
Summary:
This makes the UX more consistent. See the added comments for details.

The "write" path is a bit more complicated. Since the "date" can be
either one, we now detect the "intention" to change date and only
update "author date" with `--date`. The committer date can still be
changed via config, but not `-d`.

Reviewed By: zzl0

Differential Revision: D49214510

fbshipit-source-id: d133ed020bc4259b72528c4a72a3200262defa91