Added reason to `EdenApiMiss` events
Summary:
Today, `EdenApiMiss` events in scuba only tell us what was missed (blob or tree), but not why they were missed. We do have error information (capatured in an exception) that we can report. This will add that error information to the `EdenApaMiss` events.

As a result, our scuba usage will surge again. To remediate this, another diff will update configerator to adjust the retention of just this new column to 1 day or 100GB whichever is hit first.

See D49212103 for related configerator changes.

Reviewed By: MichaelCuevas

Differential Revision: D49211246

fbshipit-source-id: a075eb2e968e00ded206ae9c978fe58efd359101