Custom split range selector
Summary:
Replace the split range dropdowns with a new popup that lets you just click or click and drag to select commits in the stack.

I think this makes it easier to think about, easier/faster to change, and easy to visualize

Probably still some stuff we should tweak here. Maybe we should just render this selector instead of the empty state when the range doesn't exist?

I also foresee confusion when we add the right click to open split directly feature, since your stack won't be visible in that case AFAIK...

Reviewed By: quark-zju

Differential Revision: D49238813

fbshipit-source-id: 8a9c0138e6406e1611423f470601b21af3749d03