tests: silence ignored exceptions
Summary:
One of the tests is failing on Windows like:

  --- test-pull.t
  +++ test-pull.t.err
  @@ -50,6 +50,11 @@
     $ hg pull -r 'xxxxxxxxxxxxxxxxxxxy'
     pulling from ssh://user@dummy/test (glob)
     abort: unknown revision 'xxxxxxxxxxxxxxxxxxxy'!
  +  Exception ignored in: <function Popen.__del__ at 0x0000021555DF2790>\r (esc)
  +  Traceback (most recent call last):\r (esc)
  +    File "subprocess.py", line 1055, in __del__\r (esc)
  +    File "subprocess.py", line 1457, in _internal_poll\r (esc)
  +  OSError: [WinError 6] The handle is invalid\r (esc)
     [255]
     $ hg pull -r 'xxxxxxxxxxxxxxxxxx y'
     pulling from ssh://user@dummy/test (glob)

While it might be nice to figure out what's wrong exactly and fix it. It's
generally annoying to deal with `__del__` issues. For example, in this case the
"handle" might be GC-ed by something else... and that is not easy to debug.

For now, let's just silence related error messages in tests, as a workaround.

Reviewed By: zzl0

Differential Revision: D49258777

fbshipit-source-id: 954a330152a46bb7ad69764d2e78fd135f67e23d