Core derivation logic for GitDeltaManifest
Summary: With all the building blocks in place, it is time to finally implement the core derivation logic for `GitDeltaManifest`. This diff is necessarily large since it is not possible to break up the derivation into more pieces. The next diff will contain unit tests to validate that deriving the delta manifest doesn't lead to a runtime error. Additionally, it would validate that the manifest includes the expected entries. Testing beyond that would require E2E integration test which would take a while.

Differential Revision: D48783515

fbshipit-source-id: fa71b41d21ffd384e4be2810ae94a3c3e713468a