Implement helper methods for GitDeltaManifest
Summary: Since `GitDeltaManifest` is a `Manifest` type, usually it makes sense to implement the `AsyncManifest` trait so we can use `ManifestOps` methods to access its elements. However `GitDeltaManifest` is not a regular manifest that is structured like a tree and is instead a flat manifest which is structured like a list. Hence, instead of implementing the `Manifest` or `AsyncManifest` trait, the helper methods for accessing the elements are defined directly on the `GitDeltaManifest` type.

Differential Revision: D49146739

fbshipit-source-id: b8eca14ae7b5a5508c50fe7393b270989a21145b