manifest: rework merge function to use a trie merge
Summary:
Currently the merge function in the derivation code lists all subentries in all parent manifests when trying to determine what to recurse on to compute new subentries. This made sense before the creation of sharded manifests as all subentries used to always exist in a single blob, but with sharded manifests this can be very expensive as it can require accessing a large number of blobs. Moreover this wastes the opportunity to reuse shards from parent manifests that do not conflict with any other manifest or any of the changes.
This diff refactors this logic to use a trie merge algorithm that recursively expands the subentries in the parent manifests and in the changes byte by byte, reusing parent maps when possible. This is all done generically through the TrieMapOps trait

Differential Revision: D48954396

fbshipit-source-id: ab72e6a14d9bccfa5f25c48a660a04c8a6335382