rage: try to fix fatal timeout errors
Summary:
Recently we've been seeing a lot of rage collection errors. I don't know exactly what is happening, but typically we see a TimedOut exception that has bubbled to the main thread, and/or an "abort: IO::main() is not available (dropped)" error.

We do something unsafe where we run sl commands in-process in a thread (the thread gives us the ability to join() with a timeout). If the thread hasn't finished in time, we raise an exception within the thread (and don't wait any further). It is not hard to imagine this causing problems when sl is fundamentally not thread safe and has various global state.

I made three changes to try to fix/avoid the errors:
1. Ignore the TimedOut exception at the top level to be sure it never bubbles up and exits us.
2. When a thread doesn't finish in time and we inject an exception, give it a few seconds to exit naturally. This could let it clean up normally in some cases.
3. In bindings.commands.run, only restore the global IO back if our current IO override is still active. In other words, if another thread has changed the global IO since we started, don't mess with it.

Reviewed By: quark-zju

Differential Revision: D49289240

fbshipit-source-id: c6d8471cb1df0239cacaeb2e1a3d12b929bc36ef