Use a hook for submit confirmation
Summary:
If we use a hook for the submit stack confirmation modal, we can use a recoil callback to get all the state we need.

We actually had a bug before, where if you disable the confirmation modal, the submit button would have no effect, since now it returns null. We actually need it to return the same information by looking up recoil state for the draft checkbox.

Reviewed By: quark-zju

Differential Revision: D49333423

fbshipit-source-id: 3a398667ed986eb2ba5122574c1dd22179727606