Add settings checkbox for stack submit confirmation
Summary:
Update the settings dropdown to include the stack submit confirmation.

This mirrors the "don't show again" in the confirmation modal.

Of course, we need this to be accessible outside the modal because once you check it and close the modal, you won't see the modal again!

Also, we have a few settings now for code review, let's lump them together so it's easier to read and we don't take as much space.

Reviewed By: quark-zju

Differential Revision: D49330148

fbshipit-source-id: f907f8c7e3c1d59cb76bb223851b784bace93c58