add update message field to commit info view
Summary:
We now support an update message field when you bulk submit your stack, but we should also let you do this in the commit info view.

I don't love that we need to take up an additional line of space for this, it kind of gets in the way. But I think it's better to have this feature than not.

Reviewed By: sggutier

Differential Revision: D49338613

fbshipit-source-id: cbc26a3366d2419dda1380eb11b0cc18cf2fa578