Persist and respect confirmation interstitial setting
Summary:
Actually store the config state for whether you want to see the stack submit confirmation modal.

We do this with our normal persist effect.

I was a little back and forth on whether this should be the positive "show confirmation" or the negative "don't show confirmation". I opted for the setting to be positive, but show the negative in the modal. Somehow that makes more sense to me.

Reviewed By: quark-zju

Differential Revision: D49330149

fbshipit-source-id: 7601aa3b979a460ffd80704dbb8010197690bbe3