Add left/right arrows to navigate pages of uncommitted changes
Summary:
See [#727](https://github.com/facebook/sapling/issues/727)

We want to be able to navigate through long lists of changes.

Sometimes, we have the full list available and we just need to paginate it. (uncommitted changes is like this)

Sometimes, we don't have the full list on the client, and so we want the banner but no pagination buttons. (committed changed files list is like this)

We can detect this by comparing if totalFiles and filesSubset.length are equal. This gets around us needing to pass in more complicated `pageNum`+`nextPage` props from the parent—we can handle it in the child.

Reviewed By: quark-zju

Differential Revision: D49465010

fbshipit-source-id: 8ba3861b55ecb0deb38b9f28af5dbdf4fc80809c