Fix comparison view not having any margins
Summary:
The comparison view was getting sized to 100vw / vh, so there was no hint that it's a modal on top of ISL.

It makes it feel kind of claustrophobic. This was due to the width/height being set as fit-content with max-width 100vw etc. Let's instead not set the width on the modal itself so it uses the sizing of the modal-contents child (which as margin to control itself properly)

This also means the loading spinner won't be tiny, it'll take the same space. Same for the empty state. This makes it feel less jumpy.

Reviewed By: quark-zju

Differential Revision: D49474754

fbshipit-source-id: eac964b38a8ce8a4c5c78c9c5ad2528713148255