Hold alt key to show full file paths
Summary:
Small quality of life feature: if you hold the alt key, render files with their full file path.

This is easier than changing the render type to "full paths", if you just want to take a quick peak at where a file comes from.

We had this in the old ISL and it's kinda useful I think

Reviewed By: quark-zju

Differential Revision: D49478934

fbshipit-source-id: 8c189152339c1dd4471e561a839d141c60f07e48