Fix "Hold shift to drop into editor" on drag & drop
Summary:
Annoying issue in vscode ISL, if you tried to drag an image to upload it, you'd get a "hold shift to drop into editor" prompt. You'd have to be holding shift already when dragging to be able to drag it into the text area.

We just need a root level event.preventDefault() and event.stopPropagation() to remove this behavior.

Reviewed By: quark-zju

Differential Revision: D49474302

fbshipit-source-id: 53589c6081665bdf025de5f92484cf677840cc41