Fix tooltips breaking on hot reload
Summary:
This is a dev-only issue, but it's sometimes annoying. We need a root component in the tree that we can portal our tooltips onto (to render them globally unrelated to the normal dom).

We cache this root component that we lookup in the tooltip creator.

In certain hot reloads, the tooltip container could have its DOM node change. This makes the cached container now invalid/not rendered (yet still technically a non-null DOM node). Thus all the tooltips are rendered to an unattached DOM node so they don't appear.

Three cleanups:
- make a real component for the root container instead of just ensuring we have the right div class name rendered at the root and then looking up it.
- use a hook to set the global variable for the root. This way, we can unset it on unmount
- if we use a real component like this, we don't need the super hacky test-only resetMemoizedTooltipContainer junk. So we can clean up a lot of tests!

Reviewed By: quark-zju

Differential Revision: D49477300

fbshipit-source-id: a51faf9a425847183c262ce9829c8c6bfa732e78