Update ClientRequestInfo
Summary: This diff updates the implementation of `ClientRequestInfo` to allow for optional `main_id` since the client will not have the `main_id` (due to difficulties in computing it at client-end) and they still need to send the `ClientRequestInfo` header. This diff also adds a `new` method for creating `ClientRequestInfo` with `None` main_id along with a randomly generated `client_correlator`. It also extends the set of possible `entry_points`.

Reviewed By: zzl0

Differential Revision: D49483588

fbshipit-source-id: 4172271d1f4da8b78c8660462c87244a9d24bc62