revisionstore: Publish filestore fetch metrics to ODS
Summary:
## Purpose

We want better visibility in EdenFS cache hit metrics.

As part of this effort, we want to create new ODS counters for different parts of the stack, to see which caches get hit.

At the bottom of the stack, edenFS may rely on the revisionstore to actually query Mononoke for blobs, LFS objects or aux data.

## What we do

Before this diff, we already collected information about cache hits, cache misses etc. on fetch operations of the filestore, but we didn't publish them to ODS.
After this diff, we publish these metrics to ODS.

## How we do it

We use `stats::dynamic_singleton_counter` to create ODS counters for each of the `FilestoreFetchMetrics` that we collect.

We use a dynamic counter so we can create the appropriate counters without having to manually list the exploded list of metrics that are already represented by the `FilestoreFetchMetrics` type.

We use a singleton counter so that the aggregation is done consistently on the Rust side of the binary. This is necessary as ODS counters rely on thread local storage to communicate between the aggregation and the stats gathering. That does not work across languages as Rust and C++ may interpret the bytes in thread local storage differently based on how each language encodes their type.

To make the singleton counter approach work, we need a `FacebookInit` object.

We get it with `assume_init`, which is the only way to get hold of the `FacebookInit` that was generated by calling `fbinit` in C++ from Rust. `FacebookInit` is not ffi-safe, so we couldn't pass it down if we wanted to.

To avoid calling `assume_init` in a context where `fbinit` was never called, we feature gate the ODS code behind an `ods` feature.
We create a new library: `revisionstore_with_ods` which uses this feature.
Only `backingstore` now depends on `revisonstore_with_ods`. `backingstore` is the entrypoint between EdenFS and the revisionstore. This means that only code from `EdenFS` which definitely called `fbinit` will exercise the code branch with `assume_init`.

This incidentally allows us to only add the `stats` and `fbinit` dependencies to `revisionstore_ods`, which means we don't affect the `revisionstore` library which is also used by Mercurial itself.

Reviewed By: MichaelCuevas

Differential Revision: D49438188

fbshipit-source-id: c44eaa46deb00f37165ab13e3525c409c09af03e