construct main_client_id
Summary:
Let's construct singular id for a client. This is intended to be equivalent to SR's client_id. We'll use it for dynamic rate limitting - we'll track usage against the main_client_id and dynamically rate limit if it goes over threshold.

For now we're inferring it just out of identities, but later one we might want to add other info like sandcastle alias. The logic of how the main_client_id is constructed is subject to change.

Reviewed By: RajivTS

Differential Revision: D49440821

fbshipit-source-id: 8f18186e28420cfd17b14afc6323440876620aef