Use latest message + unsaved edits in split state
Summary:
When we load the commit stack via exportStack, it doesn't account for the latest remote message OR any unsaved local edits.

Due to weird predecessors, we can't easily reassociate edited messages after splitting. So we show a modal to confirm if the user wants to apply their local edits. If they confirm that, then the idea is to apply the local edits when loading the split UI. This allows us to skip running metaedit for every message with changes and then waiting for that to complete.

The end result is the same, because running split will effectively apply those local edits.

Reviewed By: quark-zju

Differential Revision: D49551315

fbshipit-source-id: 22e9f57b69f7897bd0a70ed4368b42045fcd3b3e