Refactor selector for unsaved edited commit message
Summary:
When computing hasUnsavedEditedCommitMessage, we have to compute the fields that are being edited for the commit, and then basically check if any are true.

We can make a selector dedicated to the first part, which finds all fields being edited, and use it to compute the second part trivially.

This lets us reuse the fields being edited which is useful.

Reviewed By: quark-zju

Differential Revision: D49551317

fbshipit-source-id: fd945f456e1bb1362bc01e1ef067ff49393b5960