Use confirmation modal before edit stack
Summary:
Like the previous diff which shows a confirmation before split from right click, we need the same confirmation before *any* edit stack. This confirmation was already set up to handle multiple commits, but needed some tweaks now that it actually does handle it.

Technically, the only dangerous case with edits is from split making weird successors. But we don't know beforehand if you'll split when using edit stack. So we need the entire modal to handle it. I think this is generally more consistent and better anyway, it could impact fold and ensures the UI is OK to show the latest commit titles.

Reviewed By: quark-zju

Differential Revision: D49551316

fbshipit-source-id: b28bc36666c86435bd6453e7b2505737d5102894