Confirm being splitting with unsaved message edits
Summary:
Add confirmation modal before you open the split UI, to ask if the user wants to confirm their unsaved edits to the commit message.

This is only required because split can produce strange successors, which will later cause ISL to reapply edited commit messages to unrelated commits from the original. This can duplicate messages and diffs and other weird stuff.

The confirmation here lets the user chose to immediately discard or immediately save the results.

This diff adds the discard path, but see later diffs in the stack for the "save". It won't actually save just yet, it just loads the edited message into the split UI's data, so that when you DO split, it will apply the changes. This allows us to skip waiting for metaedit and doing complicated syncing.

The confirmation here also shows you which commit(s) have unsaved changes, and which fields are unsaved, like title/test plan etc. I think this will be helpful to confirm things.

I believe there's a bug (which I haven't tracked down) in message syncing that can make a commit appear edited even if it was just due a remote syncing difference. So showing the user what's changed makes it easier to think about (IMO).

Ideally I guess we'd even show the diff if you hover the fields...but I don't do that just yet.

Reviewed By: quark-zju

Differential Revision: D49551314

fbshipit-source-id: 7581726c8c18e7459319a5be9b6f8bf626169da8