Use real <Commit> in split confirm modal
Summary:
I realized we would much rather use the real Commit in the unsaved edits confirmation, instead of a bespoke rendering of commit titles.

This gives us:
- remote synced message (though not the locally unsaved edited title I guess)
- proper commit bubble
- diff status, comments, time stamp, build status, etc.
- expected responsive behavior

Note we need to hide a few things from the commit to make this looks more natural, such as the selection background and the unsaved message dot.

Reviewed By: quark-zju

Differential Revision: D49551902

fbshipit-source-id: 8af2da6a767d4811adc85c63c5fd75da8dd2ae00